Lets have some fun and use some multi methods
Feels like it cleaned up the mapping of input to operation nicely. I
wish there was a way to ignore the input.

Thinking about collapsing the game-board atoms into a dict that gets
passed around later. So the flow can just take a snapshot of the state
and move on, instead of always reaching for a global.